STENCIL-3243 Renaming stencil-build-config to build-config
