+++ bin/stencil-start
@@ -13,7 +13,7 @@ var Pkg = require('../package.json');
 var Program = require('commander');
 var Server = require('../server');
 var ThemeConfig = require('../lib/theme-config');
-var buildConfig = require('../lib/stencil-build-config');
+var buildConfig = require('../lib/build-config');
 var stencilToken = require('../server/lib/stencil-token');
 var Url = require('url');
 var Wreck = require('wreck');
