+++ lib/build-config.js
