+++ lib/build-config.spec.js
@@ -15,10 +15,10 @@ describe('stencilBuildConfig', () => {
 
     function loadModule(mockName) {
         const path = `${cwd}/test/_mocks/build-config/${mockName}`;
-        delete require.cache[require.resolve('./stencil-build-config')];
+        delete require.cache[require.resolve('./build-config')];
         sandbox.stub(process, 'cwd').returns(path);
 
-        return require('./stencil-build-config');
+        return require('./build-config');
     }
 
     lab.beforeEach(done => {
