+++ test/_mocks/build-config/noready-config/stencil.conf.js
@@ -45,8 +45,6 @@ if (process.send) {
             production();
         }
     });
-
-    // process.send('ready');
 }
 
 module.exports = { watchOptions };
